requirements.txt
amqp==5.0.6
appdirs==1.4.4
asgiref==3.3.4
attrs==20.3.0
bcrypt==3.2.0
billiard==3.6.4.0
black==21.5b0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
distro==1.5.0
Django==3.2
django-seed==0.2.2
djangorestframework==3.12.4
djangorestframework-api-key==2.0.0
docker==5.0.0
docker-compose==1.29.1
dockerpty==0.4.1
docopt==0.6.2
Faker==8.1.2
idna==2.10
iniconfig==1.1.1
jsonschema==3.2.0
kombu==5.0.2
mypy-extensions==0.4.3
packaging==20.9
paramiko==2.7.2
pathspec==0.8.1
Pillow==8.2.0
pluggy==0.13.1
prompt-toolkit==3.0.18
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pytest==6.2.4
pytest-black==0.3.12
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
texttable==1.6.3
toml==0.10.2
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.59.0
